Начинает работу над 7 разделом
