+++ js/utils.js
@@ -1,0 +1,2 @@
+export const getRandomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+export const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
